Fix Depth First Copy Collection

Depth first copy collection was implemented incorrectly. Roots and children were added to the stack first child/root first, meaning that the last child/root is the first to be popped and copied. This commit fixes this problem, and also removes some compile warnings left over from the last commit.

See merge request !3

+++ Allocators/Allocator.cpp
@@ -61,7 +61,7 @@ size_t Allocator::getOldSpaceEndHeapIndex(){
 }
 
 size_t Allocator::getSpaceToNextObject(size_t start){
-	for(int i=start; i<oldSpaceEndHeapIndex; i++){
+	for(size_t i=start; i<oldSpaceEndHeapIndex; i++){
 		if(isBitSet(i)){
 			return i-start;
 		}
@@ -70,7 +70,7 @@ size_t Allocator::getSpaceToNextObject(size_t start){
 }
 
 unsigned char *Allocator::getNextObjectAddress(size_t start){
-	for(int i=start; i<oldSpaceEndHeapIndex; i++){
+	for(size_t i=start; i<oldSpaceEndHeapIndex; i++){
 		if(isBitSet(i)){
 			return &heap[i];
 		}
